+++ ecmascript/js_proxy.cpp
@@ -863,7 +863,7 @@ JSTaggedValue JSProxy::CallInternal(EcmaRuntimeCallInfo *info)
     // 8.Let argArray be CreateArrayFromList(argumentsList).
     ObjectFactory *factory = thread->GetEcmaVM()->GetFactory();
     size_t argc = info->GetArgsNumber();
-    JSHandle<TaggedArray> taggedArray = factory->NewTaggedArray(argc);
+    JSHandle<TaggedArray> taggedArray = factory->NewTaggedArray(static_cast<uint32_t>(argc));
     for (size_t index = 0; index < argc; ++index) {
         taggedArray->Set(thread, index, info->GetCallArg(index));
     }
@@ -914,7 +914,7 @@ JSTaggedValue JSProxy::ConstructInternal(EcmaRuntimeCallInfo *info)
     // 8.Let argArray be CreateArrayFromList(argumentsList).
     ObjectFactory *factory = thread->GetEcmaVM()->GetFactory();
     size_t argc = info->GetArgsNumber();
-    JSHandle<TaggedArray> taggedArray = factory->NewTaggedArray(argc);
+    JSHandle<TaggedArray> taggedArray = factory->NewTaggedArray(static_cast<uint32_t>(argc));
     for (size_t index = 0; index < argc; ++index) {
         taggedArray->Set(thread, index, info->GetCallArg(index));
     }
